refactor: extract serialization methods and types
