+++ src/types/serialization.ts
@@ -1,0 +1,18 @@
+import { ClassPopularityIndex, IntegrationProfile, RezervoClassBase } from "@/types/rezervo";
+
+export type RezervoScheduleDTO = { [weekOffset: number]: RezervoWeekScheduleDTO };
+export type RezervoWeekScheduleDTO = RezervoDayScheduleDTO[];
+export type RezervoDayScheduleDTO = {
+    date: string;
+    classes: RezervoClassDTO[];
+};
+export type RezervoClassDTO = RezervoClassBase & {
+    startTime: string;
+    endTime: string;
+};
+
+export type IntegrationPageProps = {
+    integrationProfile: IntegrationProfile;
+    initialSchedule: RezervoScheduleDTO;
+    classPopularityIndex: ClassPopularityIndex;
+};
